proto: Make instance name more predictable for user blocks
The if/else if/else if/ ... chain of generate makes instance
name in the flat verilog very very long ... so we avoid the
'else' part and make the generate block explicitely named.

Then we need to make a bunch of empty 'if' just to get to
the final 'else' when no block matched that position.

Signed-off-by: Sylvain Munaut <tnt@246tNt.com>