py/tt: Remove the SDim concept and use align methods
We don't actually need to align on sites, we're more interested
on aligning tracks and sites. And then we only used 'sites' unit
during the main method and nowhere else.

So instead, we store and use native units everywhere and delegate
alignement to an 'align' method that takes care of things

Signed-off-by: Sylvain Munaut <tnt@246tNt.com>