ol2/tt_{ctrl,mux}: Fix potential overlap issue for diodes
So when we insert diode ourselves, the padding is zero by default
for those. But if antenna repair during GRT is enabled, the diode
padding is set to non-zero by default meaning that during the
placement check of GRT, it will potentially read the diode placement
of DPL as invalid since that placement was made with no padding ...

So we just set padding value during GRT antenna repair to 0.

Signed-off-by: Sylvain Munaut <tnt@246tNt.com>