rtl/prim_sky130: Use 'bufinv' for strong inverter
A strong inverter has a lot of gate capacitance to drive,
so use 'bufinv' which has a buffer in front of it.

It's not a problem for 'buf' primitive since inherently
you have a small inverter first then a strong one, so the
first stage isn't that hard to drive.

Signed-off-by: Sylvain Munaut <tnt@246tNt.com>