ol2: Add workaround for OpenROAD#3753
Basically when routing tie-offs we can only use one segment or we get
hit with the bug. So we just use one line on met4 across all the pins.

But we also need to tweak the mux to make sure that line will not
conflict with any met4 on the inside of the mux, hence the obstruction

To be removed when this is fixed in OpenROAD.

Signed-off-by: Sylvain Munaut <tnt@246tNt.com>