ol2/tt_top: Setup STA for top-level
This setups the STA for the top-level.

The special thing is that we want to analyze some
"internal" path, to/from the IO from/to the user
modules.

To achieve that, we replace all the user modules
with a fake template that contain buffers for all
user modules IO and then we can reference those
buffers in the STA.

Signed-off-by: Sylvain Munaut <tnt@246tNt.com>