ol2/tt_block: Generate templates for power gated user blocks
When an user block is meant to be power gated, a small slice on
the left is stolen to be used by the power switch. So we need new
templates that are a bit narrower and that also have their pin
position adjusted so that when the block is instanciated a bit
offset, they still line up perfectly with the mux.

Signed-off-by: Sylvain Munaut <tnt@246tNt.com>