ol2/tt_block: Make sure the template is really empty
Until now we needed "something" in the block to make it go through
the flow. And so there was an assign to zero. But this created a spurious
"tie" block in the template.

Using (* noblackbox *) we can get yosys to accept an empty design.
This requires a fix in OL2 too to prevent failing on 'show' command.

Signed-off-by: Sylvain Munaut <tnt@246tNt.com>