Reorganize block addressing to split physical and logical block position
Previously we were trying to maintain some 'direct' link between
the block logical ID (i.e. the address to activate it) and its
position on the chip.

This yielded some weirdness in the code that it was sometime a bit
hard to keep track of and for the upcoming split of the spine, this
made things even harder.

So instead we now have two distinct "grids":

 - Physical grid : which gives the X/Y position of the block (or at
   least the position of the connection block in the case of large
   blocks).

   This starts at X=0,Y=0 lower left corner and then goes linearly
   in both direction.

 - Logical grid : This is 10 bits and contains the Mux ID (5 MSBs)
   and the Block ID (5 LSBs).

This change is rather "invasive" as this knowledge is sort of used
at a lot of places and they all need to be updated at once.

Signed-off-by: Sylvain Munaut <tnt@246tNt.com>