ol2/tt_{ctrl,mux}: Remove custom no_synth list
We used to need it, but now we're not using any forbidden cell
and we get better results by blacklisting all _1 cells since we
have no resizer.

We leave the custom lists in tt_top because we need to be able
to reference some cells. There is no "synthesis" really in tt_top so
it doesn't really affect anything from the generated logic anyway.

Signed-off-by: Sylvain Munaut <tnt@246tNt.com>