cfg,py,rtl: Include 'tech' and PDK in the config and check when loading
Now there is no default config loaded if nothing is defined.

Instead it will look if there is a PDK env var, or a TT_CONFIG env var
and select a config based on that.

It also checks consistency between selected config and what the PDK env
var says (if present)

Signed-off-by: Sylvain Munaut <tnt@246tNt.com>